backend/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
cloudinary==1.44.1
dj-database-url==3.1.0
Django==6.0
django-appconf==1.2.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-db-connection-pool==1.2.6
django-filter==25.2
django-imagekit==5.0.0
djangorestframework==3.14.0
greenlet==3.3.1
gunicorn==24.1.1
idna==3.11
mysqlclient==2.2.7
packaging==26.0
pilkit==3.0
pillow==11.0.0
psycopg2-binary==2.9.11
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.46
sqlparams==6.2.0
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
whitenoise==6.11.0
